fix observations
