+++ CONVERT_SERVICE/convert_service/convert_service/urls.py
@@ -31,6 +31,6 @@ from django.conf.urls.static import static
 from django.conf import settings
 
 urlpatterns = [
-                  path('admin/', admin.site.urls),
-                  path('convert/', include('convert_app.urls')),
-              ] + static("/media", document_root=settings.MEDIA_ROOT)
+    path('admin/', admin.site.urls),
+    path('convert/', include('convert_app.urls')),
+] + static("/media", document_root=settings.MEDIA_ROOT)
